requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.2.1
lxml_html_clean==0.4.1
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
